مراجعة سبتمبر (٢): الواقعة لا تُقيَّد دفعتين، والإقفال شهادة، وقائمة الدخل بلا سحوبات المالك
- createPayment يسأل التوأم ويرمي ما لم يُقَرّ؛ recordBankPayment يتبنّى المقيَّد
  بلا حركة مخصَّصاً كان أو لا؛ claimBankTransaction يربط بشرطٍ يُعدّ أثره
  (FIN-002 FIN-008 BTN-017 SCN-007).
- الإقفال: غياب الكشف والمعادلة المجهولة مانعان، ورصيدا الشهر من الاستيراد
  أو نموذج /close، ورابطٌ لكلّ بند، والشهر المقفل لا يُكتب فيه في الخدمات
  وفي القاعدة (هجرة 028)، وإعادة الفتح للمالك (FIN-007 SCN-015 BTN-015 SCN-009 SEC-015 SCN-019).
- قائمة الدخل تستبعد التحويل الشخصيّ وشراء البضاعة بأسماء عربية (FIN-005 BRW-002)،
  والمصروف يتبع التصنيف (FIN-006)، والمورّد يُقارَن بكشفه في تاريخه (FIN-004 GAP-014).
- سطر كشفٍ لم يُقرأ يُعلَن (FIN-010)، والتنبيهات لا تتضاعف (BTN-011)،
  وdetectAnomalies موصولة (FIN-012)، وتوقيت الرياض (FIN-009)، وbigint (FIN-013)،
  والأرقام العربية في الرقم الضريبي (FIN-014)، والتحجيم بأعداد صحيحة (FIN-015).
- أسماء أفعال التدقيق صادقة وبالهللات، وتراجعٌ عن «ليست سداداً» (BTN-031)،
  ورسالة حجبٍ عربية (BTN-008).
- db:migrate: لا إعادة صامتة وقفلٌ استشاريّ (CODE-002 OPS-011). db:verify يطابق
  القيد: ١٤ من ١٤ (CODE-001). وهجرة 029 للفهارس (OPS-016).

١٬٧١٠ اختباراً · typecheck وlint وbuild نظيفة · الهجرتان مطبَّقتان.

+++ src/app/suppliers/page.tsx
@@ -34,14 +34,14 @@ export default async function SuppliersPage() {
         select count(*)::int from invoices i where i.supplier_id = suppliers.id
       )`,
       billedMinor: sql<number>`(
-        select coalesce(sum(i.total_minor), 0)::int from invoices i where i.supplier_id = suppliers.id
+        select coalesce(sum(i.total_minor), 0)::bigint from invoices i where i.supplier_id = suppliers.id
       )`,
       /*
         كلُّ ما دُفع له فعلاً — لا ما خُصّص على فواتيره وحده. مالٌ دُفع ولم
         يُخصّص كان لا يُرى، فيبدو المورّد مديناً وقد سُدّد.
       */
       paidMinor: sql<number>`(
-        select coalesce(sum(p.amount_minor - p.fee_minor), 0)::int
+        select coalesce(sum(p.amount_minor - p.fee_minor), 0)::bigint
         from payments p
         where p.supplier_id = suppliers.id and p.status not in ('REVERSED', 'VOID')
       )`,
